requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
docopt==0.6.2
factory-boy==3.2.0
Faker==6.6.2
filelock==3.0.12
flake8==3.9.0
gunicorn==20.0.4
idna==2.10
mccabe==0.6.1
Pillow==8.1.2
psycopg2==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
virtualenv==20.4.2
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
